+++ node.gyp
@@ -178,7 +178,6 @@
       'lib/internal/repl/utils.js',
       'lib/internal/socket_list.js',
       'lib/internal/test/binding.js',
-      'lib/internal/test/heap.js',
       'lib/internal/timers.js',
       'lib/internal/tls.js',
       'lib/internal/trace_events_async_hooks.js',
