+++ test/common/heap.js
@@ -3,14 +3,88 @@
 const assert = require('assert');
 const util = require('util');
 
-let internalTestHeap;
+let internalBinding;
 try {
-  internalTestHeap = require('internal/test/heap');
+  internalBinding = require('internal/test/binding').internalBinding;
 } catch (e) {
   console.log('using `test/common/heap.js` requires `--expose-internals`');
   throw e;
 }
-const { createJSHeapSnapshot, buildEmbedderGraph } = internalTestHeap;
+
+const { buildEmbedderGraph } = internalBinding('heap_utils');
+const { getHeapSnapshot } = require('v8');
+
+function createJSHeapSnapshot() {
+  const stream = getHeapSnapshot();
+  stream.pause();
+  const dump = JSON.parse(stream.read());
+  const meta = dump.snapshot.meta;
+
+  const nodes =
+    readHeapInfo(dump.nodes, meta.node_fields, meta.node_types, dump.strings);
+  const edges =
+    readHeapInfo(dump.edges, meta.edge_fields, meta.edge_types, dump.strings);
+
+  for (const node of nodes) {
+    node.incomingEdges = [];
+    node.outgoingEdges = [];
+  }
+
+  let fromNodeIndex = 0;
+  let edgeIndex = 0;
+  for (const { type, name_or_index, to_node } of edges) {
+    while (edgeIndex === nodes[fromNodeIndex].edge_count) {
+      edgeIndex = 0;
+      fromNodeIndex++;
+    }
+    const toNode = nodes[to_node / meta.node_fields.length];
+    const fromNode = nodes[fromNodeIndex];
+    const edge = {
+      type,
+      to: toNode,
+      from: fromNode,
+      name: typeof name_or_index === 'string' ? name_or_index : null
+    };
+    toNode.incomingEdges.push(edge);
+    fromNode.outgoingEdges.push(edge);
+    edgeIndex++;
+  }
+
+  for (const node of nodes) {
+    assert.strictEqual(node.edge_count, node.outgoingEdges.length,
+                       `${node.edge_count} !== ${node.outgoingEdges.length}`);
+  }
+  return nodes;
+}
+
+function readHeapInfo(raw, fields, types, strings) {
+  const items = [];
+
+  for (let i = 0; i < raw.length; i += fields.length) {
+    const item = {};
+    for (let j = 0; j < fields.length; j++) {
+      const name = fields[j];
+      let type = types[j];
+      if (Array.isArray(type)) {
+        item[name] = type[raw[i + j]];
+      } else if (name === 'name_or_index') {  // type === 'string_or_number'
+        if (item.type === 'element' || item.type === 'hidden')
+          type = 'number';
+        else
+          type = 'string';
+      }
+
+      if (type === 'string') {
+        item[name] = strings[raw[i + j]];
+      } else if (type === 'number' || type === 'node') {
+        item[name] = raw[i + j];
+      }
+    }
+    items.push(item);
+  }
+
+  return items;
+}
 
 function inspectNode(snapshot) {
   return util.inspect(snapshot, { depth: 4 });
